FST 84 restructure
