+++ src/features/integration/types/Property.ts
@@ -1,0 +1,10 @@
+export interface Property {
+    order: number;
+    key: string;
+    source: string;
+}
+
+export interface PropertyString {
+    value?: string;
+    properties?: Property[]
+}
