+++ src/features/integration/types/ValueBuilderStrategy.enum.ts
@@ -1,0 +1,4 @@
+export enum VALUE_BUILDER_STRATEGY {
+    FIXED_ARCHIVE_CODE_VALUE,
+    COMBINE_STRING_VALUE
+}
