+++ src/features/integration/util/util.ts
@@ -1,20 +1,5 @@
 import {ValueBuilder} from "../types/Field";
-
-export interface property {
-    order: number;
-    key: string;
-    source: string;
-}
-
-export interface propertyString {
-    value: string;
-    properties?: property[]
-}
-
-export enum VALUE_BUILDER_STRATEGY {
-    FIXED_ARCHIVE_CODE_VALUE,
-    COMBINE_STRING_VALUE
-}
+import {Property} from "../types/Property";
 
 /* TODO: Remove empty check after validation ensures no empty fields */
 export function createValueBuilder(inputString: any): ValueBuilder {
@@ -27,7 +12,7 @@ export function createValueBuilder(inputString: any): ValueBuilder {
     }
     return {
         value: stringWithReplacedTags,
-        properties: foundTags?.map((tag: any, index: number): property => {
+        properties: foundTags?.map((tag: any, index: number): Property => {
             return {
                 key: tag,
                 order: index,
